move formfield to common
