+++ src/components/IntegrationDetailDevelop/ConfigureRunnerModal/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { MenuItem, Select } from '@material-ui/core';
-import Textarea from '../../FormFields/Textarea';
+import Textarea from '../../common/FormFields/Textarea';
 
 export const Card = styled.div<{ open: boolean }>`
   position: absolute;
