+++ src/components/common/FormFields/TextField/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { TextField } from '@material-ui/core';
-import { inputBlueMixin, withError } from '../../globalStyle';
+import { inputBlueMixin, withError } from '../../../globalStyle';
 
 export const MUITextField = styled(TextField)<{ $fieldVariant: 'default' | 'customBlue'; $hasError?: boolean }>`
   ${(props) =>
